requirements.txt
python-telegram-bot[job-queue]>=20.0
requests>=2.25.0
APScheduler>=3.6.0
